rearrange file system
