+++ backend/src/ee/services/pam-account-template/pam-account-template-schemas.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { PamRecordingStorageBackend } from "../pam-session-recording-storage/pam-session-recording-storage-enums";
+import { PamRecordingStorageBackend } from "../pam-session-recording/pam-recording-enums";
 
 export const PamPasswordConstraintsSchema = z.object({
   minLength: z.number().int().min(1).max(256),
