--- backend/src/ee/services/pam-session/pam-session-enums.ts
@@ -1,6 +1,0 @@
-export enum PamSessionStatus {
-  Starting = "starting",
-  Active = "active",
-  Ended = "ended",
-  Terminated = "terminated"
-}
