+++ backend/src/ee/services/pam/pam-enums.ts
@@ -22,3 +22,10 @@ export enum PamProductRole {
   Admin = "admin",
   Member = "member"
 }
+
+export enum PamSessionStatus {
+  Starting = "starting",
+  Active = "active",
+  Ended = "ended",
+  Terminated = "terminated"
+}
